An end of quiz leaderboard is now displayed when the active quiz finishes - displaying the team name and score of each team that participated in the quiz in order of descending score.
The currently logged-in user's score is also highlighted on this scoreboard.

Also, the getUserData function in the database class has been split into two functions: getUserData and getLoggedInUserData. This was done so that a CurrentUser object can be returned given a specific user id rather than just the logged-in user's id.