Users that win the pub quiz now have one win added to their user document in the Users collection (if the quiz does not end early).
Also, the algorithm that determines the values of the three highest scores on the End of Quiz Leaderboard has been amended as previously if the second and third highest scores were equal the users in second place would have their rows coloured bronze instead of silver.

Also, logged-in user's can now access the End of Quiz Leaderboard if they leave the active quiz screen and join back - as long as a quiz admin does not tap the button that ends the quiz on the Quiz Control Widget.